requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11
django-cors-headers==2.4.0
django-extensions==2.1.0
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.7
krakenio==0.0.2
Markdown==2.6.11
Pillow==5.2.0
PyJWT==1.6.4
python-resize-image==1.1.12
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
